+++ lib/util/documents.ts
@@ -1,0 +1,16 @@
+import { MutableStringMap } from '@ts-common/string-map'
+import { SwaggerObject } from 'yasway'
+
+export type DocCache = MutableStringMap<Promise<SwaggerObject>>
+
+/*
+ * Caches the json docs that were successfully parsed by parseJson().
+ * This avoids, fetching them again.
+ * key: docPath
+ * value: parsed doc in JSON format
+ */
+export let docCache: DocCache = {}
+
+export function clearCache(): void {
+  docCache = {}
+}
